Update for Jetpack Compose compatibility
Add dependencies for Jetpack Compose and change MainActivity to use a
Compose Text to display a message on the screen.

One of the main issues in getting this to work was the version set for
the Kotlin Gradle plugin. When set to:
  const val kotlin_gradle_plugin = "1.6.21"
in Dependencies.kt it would not build. To get it to build had to change
this to:
  const val kotlin_gradle_plugin = "1.6.10"

Not sure if all the dependencies are required, and they should also be
added to the Dependencies.kt file - to be consistent with other
dependencies.